intent-engine/requirements.txt
# Core dependencies
faster-whisper>=1.0.0
ollama>=0.1.0
paho-mqtt>=1.6.0
requests>=2.31.0
numpy>=1.24.0
scipy>=1.11.0
tenacity>=8.2.3

# Audio processing
ffmpeg-python>=0.2.0
soundfile>=0.12.0
librosa>=0.10.0

# TTS options (install based on your choice)
TTS>=0.22.0  # Coqui TTS (includes XTTS)
piper-tts>=1.2.0  # Piper
# kokoro-onnx  # Uncomment if using Kokoro

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
PyYAML>=6.0.1

# NOTE: Do not add pyunifiprotect here.
# It was removed from PyPI and will break builds.

# Optional: Web UI dependencies
flask>=3.0.0
flask-cors>=4.0.0
